1165 - user error at remove scheduled filter

+++ module/core/src/main/java/ru/protei/portal/core/service/IssueFilterServiceImpl.java
@@ -5,11 +5,11 @@ import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import ru.protei.portal.api.struct.Result;
 import ru.protei.portal.core.model.dao.CaseFilterDAO;
+import ru.protei.portal.core.model.dao.PersonCaseFilterDAO;
 import ru.protei.portal.core.model.dict.En_CaseFilterType;
 import ru.protei.portal.core.model.dict.En_Privilege;
 import ru.protei.portal.core.model.dict.En_ResultStatus;
 import ru.protei.portal.core.model.ent.*;
-import ru.protei.portal.core.model.helper.CollectionUtils;
 import ru.protei.portal.core.model.helper.HelperFunc;
 import ru.protei.portal.core.model.query.CaseQuery;
 import ru.protei.portal.core.model.query.CaseTagQuery;
@@ -36,6 +36,8 @@ public class IssueFilterServiceImpl implements IssueFilterService {
 
     @Autowired
     CaseFilterDAO caseFilterDAO;
+    @Autowired
+    PersonCaseFilterDAO personCaseFilterDAO;
     @Autowired
     AuthService authService;
     @Autowired
@@ -168,8 +170,12 @@ public class IssueFilterServiceImpl implements IssueFilterService {
 
         log.debug( "removeIssueFilter(): id={} ", id );
 
+        if (personCaseFilterDAO.isUsed(id)) {
+            return error(En_ResultStatus.ISSUE_FILTER_IS_USED);
+        }
+
         if ( caseFilterDAO.removeByKey( id ) ) {
-            return ok(true );
+            return ok(true);
         }
 
         return error(En_ResultStatus.INTERNAL_ERROR );
